sql: disallow tenants from setting zone configs by default
Previously, `sql.zone_configs.allow_for_secondary_tenant.enabled` was a
regular old cluster setting that tenants could control. Now that we have
a notion of tenant read-only setting, this feels like a great candidate
for it. In addition to making this switch, this patch also changes this
setting's default value to be false. See the linked issue for the
motivation around why.

The change to switch this setting to be tenant read-only necessitates
changes in how we run logic tests. This is because logic tests that
run as secondary tenants can no longer change this cluster setting
in the test file. We introduce a new
`tenant-cluster-setting-override-opt` logic test directive with an
option for `allow-zone-configs-for-secondary-tenants` to get this
working. When configured, the host alters this cluster setting
during test setup. This can later be extended for other read-only
settings in the future.

References #77344

Release note (sql change): Changes the default value of
`sql.zone_configs.allow_for_secondary_tenant.enabled` to be false.
 Moreover, this setting is no longer settable by secondary tenants.
Instead, it's now a tenant read-only cluster setting.